Replace int id to bigint id in schema.sql

+++ src/main/resources/schema.sql
@@ -9,16 +9,16 @@ CREATE TABLE Author(
 CREATE UNIQUE INDEX author_name ON Author (name);
 
 CREATE TABLE Genre(
-  id INT PRIMARY KEY AUTO_INCREMENT,
+  id BIGINT PRIMARY KEY AUTO_INCREMENT,
   name VARCHAR(255)
 );
 CREATE UNIQUE INDEX genre_name ON Genre (name);
 
 CREATE TABLE Book (
-  id INT PRIMARY KEY AUTO_INCREMENT,
+  id BIGINT PRIMARY KEY AUTO_INCREMENT,
   name VARCHAR(255) ,
   author_id BIGINT,
-  genre_id INT,
+  genre_id BIGINT,
   CONSTRAINT name_author_genre UNIQUE (name, author_id, genre_id),
   FOREIGN KEY(author_id) REFERENCES Author(id) ON DELETE CASCADE,
   FOREIGN KEY(genre_id) REFERENCES Genre(id) ON DELETE CASCADE
